+++ src/meeting_recorder/__init__.py
@@ -1,3 +1,3 @@
 """Meeting Recorder App."""
 
-__version__ = "0.4.1"
+__version__ = "0.5.0"